feat: add ProfileImage.style
